backend/requirements.txt
Django==5.2.1
djangorestframework==3.16.0
psycopg2-binary==2.9.10
django-cors-headers==4.7.0
djangorestframework-simplejwt==5.5.0
pytest==8.4.0
pytest-django==4.11.1
dj-database-url==3.0.0
django-stubs==5.2.0
aiohttp==3.12.9
qdrant_client==1.14.2
pylint-django==2.6.1